refactor: split tunnel relay helpers
